requirements.txt
Django==1.4
django-piston==0.2.3
requests==0.13.2
